requirements.txt
pydantic>=2.7,<3
requests>=2.31,<3
python-dotenv>=1.0,<2
python-dateutil>=2.9,<3
